+++ packages/canary-integration/src/__tests__/model-instance-document.test.ts
@@ -2,7 +2,7 @@ import { jest } from '@jest/globals'
 import getPort from 'get-port'
 import { AnchorStatus, CeramicApi, CommitType, IpfsApi } from '@ceramicnetwork/common'
 import { createIPFS } from '@ceramicnetwork/ipfs-daemon'
-import { ModelInstanceDocument } from '@ceramicnetwork/stream-model-instance'
+import { ModelInstanceDocument, ModelInstanceDocumentMetadata } from '@ceramicnetwork/stream-model-instance'
 import { createCeramic } from '../create-ceramic.js'
 import { anchorUpdate } from '@ceramicnetwork/core/lib/state-management/__tests__/anchor-update'
 import { Ceramic } from '@ceramicnetwork/core'
@@ -10,6 +10,7 @@ import { CeramicDaemon, DaemonConfig } from '@ceramicnetwork/cli'
 import { CeramicClient } from '@ceramicnetwork/http-client'
 import { StreamID } from '@ceramicnetwork/streamid'
 import first from 'it-first'
+import { Model, ModelAccountRelation, ModelDefinition } from '@ceramicnetwork/stream-model'
 import {SqliteIndexApi} from "@ceramicnetwork/core/lib/indexing/sqlite/sqlite-index-api";
 import {listMidTables} from "@ceramicnetwork/core/lib/indexing/sqlite/init-tables";
 import knex, { Knex } from 'knex'
@@ -95,6 +96,26 @@ async function isPinned(ceramic: CeramicApi, streamId: StreamID): Promise<boolea
   return (await first(iterator)) == streamId.toString()
 }
 
+const MODEL_DEFINITION: ModelDefinition = {
+  name: 'MyModel',
+  accountRelation: ModelAccountRelation.LIST,
+  schema: {
+    $schema: "https://json-schema.org/draft/2020-12/schema",
+    type: "object",
+    additionalProperties: false,
+    properties: {
+      myData: {
+        type: "integer",
+        maximum: 10000,
+        minimum: 0
+      },
+    },
+    required: [
+      "myData"
+    ]
+  }
+}
+
 describe('ModelInstanceDocument API http-client tests', () => {
   jest.setTimeout(1000 * 30)
 
@@ -102,6 +123,8 @@ describe('ModelInstanceDocument API http-client tests', () => {
   let core: Ceramic
   let daemon: CeramicDaemon
   let ceramic: CeramicClient
+  let model: Model
+  let midMetadata: ModelInstanceDocumentMetadata
 
   beforeAll(async () => {
     ipfs = await createIPFS()
@@ -113,6 +136,9 @@ describe('ModelInstanceDocument API http-client tests', () => {
     await daemon.listen()
     ceramic = new CeramicClient(apiUrl)
     ceramic.setDID(core.did)
+
+    model = await Model.create(ceramic, MODEL_DEFINITION)
+    midMetadata = { model: model.id }
   }, 12000)
 
   afterAll(async () => {
@@ -122,22 +148,50 @@ describe('ModelInstanceDocument API http-client tests', () => {
     await ipfs.stop()
   })
 
-  test('Create valid doc', async () => {
-    const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, METADATA)
+  test('verifies the content against model schema when creating an MID', async () => {
+    await expect(
+      ModelInstanceDocument.create(ceramic, {}, midMetadata)
+    ).rejects.toThrow(/data must have required property 'myData'/)
+  })
 
+  test('verifies the content against model schema when updating an MID', async () => {
+    const doc = await ModelInstanceDocument.create(
+      ceramic,
+      CONTENT0,
+      midMetadata
+    )
+    expect(doc.content).toEqual(CONTENT0)
+
+    await expect(
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore
+      doc.replace({})
+    ).rejects.toThrow(/data must have required property 'myData'/)
+  })
+
+  test(`Create a valid doc`, async () => {
+    const doc = await ModelInstanceDocument.create(
+      ceramic,
+      CONTENT0,
+      midMetadata
+    )
     expect(doc.id.type).toEqual(ModelInstanceDocument.STREAM_TYPE_ID)
-    expect(JSON.stringify(doc.content)).toEqual(JSON.stringify(CONTENT0))
+    expect(doc.content).toEqual(CONTENT0)
     expect(doc.state.log.length).toEqual(1)
     expect(doc.state.log[0].type).toEqual(CommitType.GENESIS)
     expect(doc.state.anchorStatus).toEqual(AnchorStatus.PENDING)
-    expect(doc.metadata.model.toString()).toEqual(FAKE_STREAM_ID.toString())
+    expect(doc.metadata.model.toString()).toEqual(model.id.toString())
     expect(doc.metadata.unique instanceof Uint8Array).toBeTruthy()
     await expect(isPinned(ceramic, doc.id)).resolves.toBeTruthy()
   })
 
   test('Create and update doc', async () => {
-    const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, METADATA)
-    expect(JSON.stringify(doc.content)).toEqual(JSON.stringify(CONTENT0))
+    const doc = await ModelInstanceDocument.create(
+      ceramic,
+      CONTENT0,
+      midMetadata
+    )
+    expect(doc.content).toEqual(CONTENT0)
 
     // confirm stream was added to index
     /*let result: Array<any> = await dataSource.query(`SELECT * FROM mid_${FAKE_STREAM_ID}`)
@@ -150,7 +204,7 @@ describe('ModelInstanceDocument API http-client tests', () => {
 
     await doc.replace(CONTENT1)
 
-    expect(JSON.stringify(doc.content)).toEqual(JSON.stringify(CONTENT1))
+    expect(doc.content).toEqual(CONTENT1)
     expect(doc.state.log.length).toEqual(2)
     expect(doc.state.log[0].type).toEqual(CommitType.GENESIS)
     expect(doc.state.log[1].type).toEqual(CommitType.SIGNED)
@@ -164,7 +218,11 @@ describe('ModelInstanceDocument API http-client tests', () => {
   })
 
   test('Anchor genesis', async () => {
-    const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, METADATA)
+    const doc = await ModelInstanceDocument.create(
+      ceramic,
+      CONTENT0,
+      midMetadata
+    )
     expect(doc.state.anchorStatus).toEqual(AnchorStatus.PENDING)
 
     await anchorUpdate(core, doc)
@@ -174,11 +232,15 @@ describe('ModelInstanceDocument API http-client tests', () => {
     expect(doc.state.log.length).toEqual(2)
     expect(doc.state.log[0].type).toEqual(CommitType.GENESIS)
     expect(doc.state.log[1].type).toEqual(CommitType.ANCHOR)
-    expect(JSON.stringify(doc.content)).toEqual(JSON.stringify(CONTENT0))
+    expect(doc.content).toEqual(CONTENT0)
   })
 
   test('Anchor after updating', async () => {
-    const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, METADATA)
+    const doc = await ModelInstanceDocument.create(
+      ceramic,
+      CONTENT0,
+      midMetadata
+    )
     expect(doc.state.anchorStatus).toEqual(AnchorStatus.PENDING)
     await doc.replace(CONTENT1)
     expect(doc.state.anchorStatus).toEqual(AnchorStatus.PENDING)
@@ -191,11 +253,15 @@ describe('ModelInstanceDocument API http-client tests', () => {
     expect(doc.state.log[0].type).toEqual(CommitType.GENESIS)
     expect(doc.state.log[1].type).toEqual(CommitType.SIGNED)
     expect(doc.state.log[2].type).toEqual(CommitType.ANCHOR)
-    expect(JSON.stringify(doc.content)).toEqual(JSON.stringify(CONTENT1))
+    expect(doc.content).toEqual(CONTENT1)
   })
 
   test('multiple updates', async () => {
-    const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, METADATA)
+    const doc = await ModelInstanceDocument.create(
+      ceramic,
+      CONTENT0,
+      midMetadata
+    )
     await doc.replace(CONTENT1)
 
     await anchorUpdate(core, doc)
@@ -215,19 +281,19 @@ describe('ModelInstanceDocument API http-client tests', () => {
     expect(doc.state.log[3].type).toEqual(CommitType.SIGNED)
     expect(doc.state.log[4].type).toEqual(CommitType.SIGNED)
     expect(doc.state.log[5].type).toEqual(CommitType.ANCHOR)
-    expect(JSON.stringify(doc.content)).toEqual(JSON.stringify(CONTENT3))
+    expect(doc.content).toEqual(CONTENT3)
   })
 
   test('ModelInstanceDocuments are created uniquely', async () => {
-    const doc1 = await ModelInstanceDocument.create(ceramic, CONTENT0, METADATA)
-    const doc2 = await ModelInstanceDocument.create(ceramic, CONTENT0, METADATA)
+    const doc1 = await ModelInstanceDocument.create(ceramic, CONTENT0, midMetadata)
+    const doc2 = await ModelInstanceDocument.create(ceramic, CONTENT0, midMetadata)
 
     expect(doc1.id.toString()).not.toEqual(doc2.id.toString())
     expect(doc1.metadata.unique.toString()).not.toEqual(doc2.metadata.unique.toString())
   })
 
   test('Can load a stream', async () => {
-    const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, METADATA)
+    const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, midMetadata)
     await doc.replace(CONTENT1)
     await anchorUpdate(core, doc)
     await doc.sync()
@@ -240,23 +306,23 @@ describe('ModelInstanceDocument API http-client tests', () => {
   })
 
   test('create respects anchor flag', async () => {
-    const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, METADATA, { anchor: false })
+    const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, midMetadata, { anchor: false })
     expect(doc.state.anchorStatus).toEqual(AnchorStatus.NOT_REQUESTED)
   })
 
   test('create respects pin flag', async () => {
-    const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, METADATA, { pin: false })
+    const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, midMetadata, { pin: false })
     await expect(isPinned(ceramic, doc.id)).resolves.toBeFalsy()
   })
 
   test('replace respects anchor flag', async () => {
-    const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, METADATA, { anchor: false })
+    const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, midMetadata, { anchor: false })
     await doc.replace(CONTENT1, { anchor: false })
     expect(doc.state.anchorStatus).toEqual(AnchorStatus.NOT_REQUESTED)
   })
 
   test('replace respects pin flag', async () => {
-    const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, METADATA)
+    const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, midMetadata)
     await expect(isPinned(ceramic, doc.id)).resolves.toBeTruthy()
     await doc.replace(CONTENT1, { pin: false })
     await expect(isPinned(ceramic, doc.id)).resolves.toBeFalsy()
@@ -270,12 +336,17 @@ describe('ModelInstanceDocument API multi-node tests', () => {
   let ipfs1: IpfsApi
   let ceramic0: Ceramic
   let ceramic1: Ceramic
+  let model: Model
+  let midMetadata: ModelInstanceDocumentMetadata
 
   beforeAll(async () => {
     ipfs0 = await createIPFS()
     ipfs1 = await createIPFS()
     ceramic0 = await createCeramic(ipfs0)
     ceramic1 = await createCeramic(ipfs1)
+
+    model = await Model.create(ceramic0, MODEL_DEFINITION)
+    midMetadata = { model: model.id }
   }, 12000)
 
   afterAll(async () => {
@@ -286,7 +357,7 @@ describe('ModelInstanceDocument API multi-node tests', () => {
   })
 
   test('load basic doc', async () => {
-    const doc = await ModelInstanceDocument.create(ceramic0, CONTENT0, METADATA)
+    const doc = await ModelInstanceDocument.create(ceramic0, CONTENT0, midMetadata)
 
     const loaded = await ModelInstanceDocument.load(ceramic1, doc.id)
 
@@ -301,7 +372,7 @@ describe('ModelInstanceDocument API multi-node tests', () => {
   })
 
   test('load updated doc', async () => {
-    const doc = await ModelInstanceDocument.create(ceramic0, CONTENT0, METADATA)
+    const doc = await ModelInstanceDocument.create(ceramic0, CONTENT0, midMetadata)
     await doc.replace(CONTENT1)
 
     const loaded = await ModelInstanceDocument.load(ceramic1, doc.id)
@@ -317,7 +388,7 @@ describe('ModelInstanceDocument API multi-node tests', () => {
   })
 
   test('load updated and anchored doc', async () => {
-    const doc = await ModelInstanceDocument.create(ceramic0, CONTENT0, METADATA)
+    const doc = await ModelInstanceDocument.create(ceramic0, CONTENT0, midMetadata)
     await doc.replace(CONTENT1)
     await anchorUpdate(ceramic0, doc)