+++ packages/core/src/indexing/local-index-api.ts
@@ -57,4 +57,8 @@ export class LocalIndexApi implements IndexApi {
   async init(): Promise<void> {
     await this.databaseIndexApi?.init()
   }
+
+  async close(): Promise<void> {
+    await this.databaseIndexApi?.close()
+  }
 }