+++ packages/core/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.1](/compare/@ceramicnetwork/core@2.5.1-rc.0...@ceramicnetwork/core@2.5.1) (2022-06-20)
+
+**Note:** Version bump only for package @ceramicnetwork/core
+
+
+
+
+
+## [2.5.1-rc.0](/compare/@ceramicnetwork/core@2.5.0...@ceramicnetwork/core@2.5.1-rc.0) (2022-06-20)
+
+**Note:** Version bump only for package @ceramicnetwork/core
+
+
+
+
+
 # [2.5.0](/compare/@ceramicnetwork/core@2.5.0-rc.5...@ceramicnetwork/core@2.5.0) (2022-06-17)
 
 **Note:** Version bump only for package @ceramicnetwork/core