+++ packages/ipfs-daemon/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.8](/compare/@ceramicnetwork/ipfs-daemon@2.0.8-rc.0...@ceramicnetwork/ipfs-daemon@2.0.8) (2022-06-20)
+
+**Note:** Version bump only for package @ceramicnetwork/ipfs-daemon
+
+
+
+
+
+## [2.0.8-rc.0](/compare/@ceramicnetwork/ipfs-daemon@2.0.7...@ceramicnetwork/ipfs-daemon@2.0.8-rc.0) (2022-06-20)
+
+**Note:** Version bump only for package @ceramicnetwork/ipfs-daemon
+
+
+
+
+
 ## [2.0.7](/compare/@ceramicnetwork/ipfs-daemon@2.0.7-rc.3...@ceramicnetwork/ipfs-daemon@2.0.7) (2022-06-17)
 
 **Note:** Version bump only for package @ceramicnetwork/ipfs-daemon