+++ packages/pinning-ipfs-backend/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.7](/compare/@ceramicnetwork/pinning-ipfs-backend@2.0.7-rc.0...@ceramicnetwork/pinning-ipfs-backend@2.0.7) (2022-06-20)
+
+**Note:** Version bump only for package @ceramicnetwork/pinning-ipfs-backend
+
+
+
+
+
+## [2.0.7-rc.0](/compare/@ceramicnetwork/pinning-ipfs-backend@2.0.6...@ceramicnetwork/pinning-ipfs-backend@2.0.7-rc.0) (2022-06-20)
+
+**Note:** Version bump only for package @ceramicnetwork/pinning-ipfs-backend
+
+
+
+
+
 ## [2.0.6](/compare/@ceramicnetwork/pinning-ipfs-backend@2.0.6-rc.3...@ceramicnetwork/pinning-ipfs-backend@2.0.6) (2022-06-17)
 
 **Note:** Version bump only for package @ceramicnetwork/pinning-ipfs-backend