+++ packages/stream-model-instance-handler/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1](/compare/@ceramicnetwork/stream-model-instance-handler@0.2.1-rc.0...@ceramicnetwork/stream-model-instance-handler@0.2.1) (2022-06-20)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-model-instance-handler
+
+
+
+
+
+## [0.2.1-rc.0](/compare/@ceramicnetwork/stream-model-instance-handler@0.2.0...@ceramicnetwork/stream-model-instance-handler@0.2.1-rc.0) (2022-06-20)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-model-instance-handler
+
+
+
+
+
 # [0.2.0](/compare/@ceramicnetwork/stream-model-instance-handler@0.2.0-rc.3...@ceramicnetwork/stream-model-instance-handler@0.2.0) (2022-06-17)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-model-instance-handler